Brunch Refactor
The emccode.github.io website has been refactored to be a compiled site
that utilizes brunch as the compilation tool of choice. This achieves
greater content delivery speed, the ability to utilize HTML5 technology
more easily and keeping with industry standards through tools such as
bower and node, and opens up the possibility of more easily integrating
a framework such as AngularJS with or without React.